refactor: create spearate folder for hooks
