+++ src/hooks/use-on-click-outside.js
@@ -1,0 +1,20 @@
+import { useEffect } from "react";
+
+// https://css-tricks.com/hamburger-menu-with-a-side-of-react-hooks-and-styled-components/
+export default function useOnClickOutside(ref, handler) {
+  useEffect(() => {
+    const listener = (event) => {
+      if (!ref.current || ref.current.contains(event.target)) {
+        return;
+      }
+      handler(event);
+    };
+    document.addEventListener("mousedown", listener);
+    document.addEventListener("touchstart", listener);
+
+    return () => {
+      document.removeEventListener("mousedown", listener);
+      document.removeEventListener("touchstart", listener);
+    };
+  }, [ref, handler]);
+}
